requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
autopep8==2.3.2
blinker==1.9.0
bottle==0.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-core==0.3.58
langchain-text-splitters==0.3.8
langsmith==0.3.42
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.35.0
Nuitka==2.6.9
numpy==2.2.5
ordered-set==4.1.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==10.4.0
plotly==5.24.1
protobuf==5.29.4
proxy_tools==0.1.0
psutil==7.0.0
pyarrow==19.0.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2024.2
pywebview==5.4
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
shiboken6==6.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit==1.45.1
streamlit-option-menu==0.3.13
tenacity==8.5.0
toml==0.10.2
tornado==6.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
waitress==3.0.2
watchdog==4.0.2
zstandard==0.23.0
reportlab
